+++ R/addBinaries.R
@@ -57,7 +57,7 @@ addBinaries <- function(pps, folder=NULL, verbose=TRUE) {
         message('No folders to add binaries from.')
         return(pps)
     }
-    folder <- normalizePath(folder)
+    folder <- normalizePath(folder, winslash = '/')
     pps@binaries$folder <- unique(c(pps@binaries$folder, folder))
     if(is.null(binList) ||
        length(binList) == 0) {