+++ tests/testthat/test-workStudy.R
@@ -107,8 +107,8 @@ test_that('Test working with AcousticStudy object', {
     # test add recordings
     recs <- system.file('extdata', 'Recordings', package='PAMpal')
     exData <- addRecordings(exData, folder = recs, log=FALSE, progress=FALSE)
-    expect_identical(normalizePath(files(exData)$recordings$file),
-                     normalizePath(list.files(recs, full.names = TRUE)))
+    expect_identical(normalizePath(files(exData)$recordings$file, winslash = '/'),
+                     normalizePath(list.files(recs, full.names = TRUE), winslash = '/'))
     expect_warning(warnRec <- addRecordings(exData, folder = 'DNE', log=FALSE, progress=FALSE))
 
     # test warning access from recorder warning